[FIX] cleaning code

+++ api/src/main/java/com/api/bank/model/entity/QrCheck.java
@@ -7,7 +7,6 @@ import javax.persistence.*;
 import java.util.Date;
 import java.util.List;
 
-
 @Entity
 @Inheritance(strategy = InheritanceType.SINGLE_TABLE)
 @Getter
@@ -16,7 +15,6 @@ public class QrCheck extends Base {
 
     @Column(nullable = false, unique = true)
     private String checkToken;
-
     private double soldAmount;
 
     @Column(nullable = false)
@@ -27,7 +25,7 @@ public class QrCheck extends Base {
 
     public QrCheck() {}
 
-    public QrCheck(float soldAmount, int nbDayOfValidity) {
+    public QrCheck(double soldAmount, int nbDayOfValidity) {
         super();
         this.soldAmount = soldAmount;
         this.nbDayOfValidity = nbDayOfValidity;
